backend/requirements.txt
Django==5.0.3
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
django-cors-headers==4.3.1
django-filter==24.2
psycopg2-binary==2.9.9
python-decouple==3.8
gunicorn==22.0.0
whitenoise==6.6.0
pytest==8.1.1
pytest-django==4.8.0
pytest-cov==5.0.0
factory-boy==3.3.0
Faker==24.4.0
